complex for windows to kill a thread

+++ src/cpyquickhelper/parallel/threader.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
-int threader_kill_thread(int pid);
+int threader_kill_thread(unsigned long pid);
